requirements.txt
pyodbc
azure-identity
openai
